# BUG: Fix not to reindex on non-Categorical groups (GH9049)