closes #9049.
closes #9344 

*self._was_factor* is not appropriate to judge whether grouper is Categorical or not, because it can be "True" when we groupby indices (not columns). So, I added another flag *self._is_categorical* to judge Categorical state.

Also, I added a GroupBy test for MultiIndexed data, which was failed before this fix.