@ledmonster 2015-01-04
According to 8d2c2a8, original `_was_factor` seems not meaning "Categorical" but "level based grouping". Probably, ea1186d confused the meaning of `_was_factor`. How about renaming `_was_factor` to `_grouping_by_level` and to use both this and `_is_categorical` ?