@jreback 2015-01-04
maybe better to create a new flag maybe
_grouping_typ defaulting to None thrn set to a string (level or categorical)
it's internal so ok to change this