@ledmonster 2015-01-10
Improved by using ``_grouping_type`` instead of two boolean flags.