@jreback 2015-01-10
pls add a release note (bug fix section) referencing the original issue