@ledmonster 2015-01-11
Added a comment and release note, and squashed changes.