@jreback 2015-01-18
@ledmonster pls rebase this when you can and see if you can address the comments above.