@ledmonster 2015-01-19
Rebased and refactored ``Grouping`` class for ``Categorical`` grouper.

- evaluate ``_labels`` and ``_group_index`` instead of using ``_grouping_type`` flag in ``_make_labels``.
- stop to convert Categorical ``self.grouper`` to ndarray. I checked some logics using ``self.grouper``, but it seems to work fine without converting ``Categorical`` one to ``ndarray``. (by modifying L1948 a little)
- use ``cache_readonly`` instead of ``property`` for ``groups`` property.
- add tests for datetime Categorical grouper.