@jreback 2015-01-22
this looks pretty good - cleaned up s lot of older code

can u run s perf test (the vbench suite)
and report of anythjng anomalous comes up