@ledmonster 2015-01-22
I tried ```./test_perf.sh -b master -t HEAD```, but it took too much machine resources. So, I'll try again tonight while I'm in bed, with ``-r groupby`` option.

BTW, I got some problems on running perf test, and made Pull #9332 for them.