@ledmonster 2015-01-24
My Air Mac freezes by running ``./test_perf.sh -b master -t HEAD -r groupby``, so I'll try it on EC2 instance later.